requirements.txt
Django~=2.2.0
django-cors-headers~=3.0.1
djangorestframework~=3.9.4
djangorestframework-simplejwt~=4.3.0
